requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-admin==2.0.2
django-excel-response2==3.0.5
django-phonenumber-field==7.0.2
django-rest-framework==0.1.0
django-six==1.0.5
djangorestframework==3.14.0
environ==1.0
excel-base==1.0.4
idna==3.4
isoweek==1.3.3
phonenumbers==8.13.9
pycodestyle==2.10.0
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
screen==1.0.1
six==1.16.0
sqlparse==0.4.3
TimeConvert==3.0.13
tomli==2.0.1
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
xlwt==1.3.0
